requirements.txt
sentry_sdk==0.14.2
SQLAlchemy==1.3.12
numpy==1.17.4
Flask==1.1.2
requests==2.22.0
marshmallow_dataclass==8.3.2
Flask_Cors==3.0.8
PyJWT==2.1.0
stripe
gunicorn
psycopg2